feat: improve UI for data app connections
